exercise 4.4: Reschedule ProgramSlot
Using DateInterval is much simpler, but also revealed a nasty error in
previous version of tests:

ProgramSlots expected after the reschedule did not have their durations
in sync with the original slots - original ones were lasting for 1 hour,
but the after the rescheduling the expectations was that they last for
two hours.

This lead to a bug in implementation that tried to compensate for the
diff.

This could probably be fixed easily by updating the tests to be
realistic, then refactoring the implementation.