exercise 4.5: Reschedule Meeting with multiple ProgramSlots
Changed the Meeting API a bit: instead of rescheduling the Meeting with
a DateInterval, we could reschedule it for a future date. This way we
can calculate the diff between the old start and the new, and use that
to reschedule the program slots within the meeting. This makes things a
lot easier, since the Meeting is the aggregate - Program and it's
ProgramSlots don't really know anything about their relative position
within the meeting and cannot reschedule themselves by themselves.

The meeting is the main orchestrator and should be tasked with providing
its internal parts with all the information they need.